requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.1
dataclasses==0.6
Django==5.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.6.1
docker==7.0.0
idna==3.6
packaging==24.0
pillow==10.2.0
PyJWT==2.8.0
pymongo==4.6.2
pytz==2024.1
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
